Bet,BetState,BetRequest

+++ rasbet-backend/src/main/java/net/java/rasbetbackend/controller/AuthController.java
@@ -67,7 +67,7 @@ public class AuthController {
                 .collect(Collectors.toList());
 
         return ResponseEntity.ok(new JwtResponse(jwt,
-                userDetails.getNif(),
+                (long) userDetails.getNif(),
                 userDetails.getUsername(),
                 userDetails.getEmail(),
                 roles));
